+++ src/splat/segtypes/common/hasm.py
@@ -1,7 +1,18 @@
+from pathlib import Path
+from typing import Optional
+
 from .asm import CommonSegAsm
+
+from ...util import options
 
 
 class CommonSegHasm(CommonSegAsm):
+    def out_path(self) -> Optional[Path]:
+        if options.opts.hasm_in_src_path:
+            return options.opts.src_path / self.dir / f"{self.name}.s"
+
+        return super().out_path()
+
     def scan(self, rom_bytes: bytes):
         if (
             self.rom_start is not None