+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # splat Release Notes
 
+### 0.21.4
+
+* New yaml option: `hasm_in_src_path`
+  * Tells splat to consider `hasm` files to be relative to `src_path` instead of `asm_path`.
+
 ### 0.21.3
+
 * Updated version graphically
 
 ### 0.21.2
+
 * Fix bugs involving segments not having proper end rom positions if followed by segments with "auto" rom addresses; splat will now skip over these properly
 
 ### 0.21.1